Add files via upload
This Machine Learning Model is to Predict the Match Outcome from the Historical Data ..  As we know predictions are for the future according to the past, So this csv file is the dataset of English Premier League season ranging from (2020 to 2021)and (2021 to 2022) and I am predicting the match outcome on future either Win or Loss and Draw. if it's a Win it is treated as 1 and if it's a loss or draw it is 0.